Fixes to support Raspberry Pi
Raspberry Pi is a armv7-a cortex-a7 with NEON. It's default OS is
Raspbian which is a Debian offshoot. From a build environment
perspective, it is almost indistiguishable from any other Debian based
Linux.

Pi supports OpenGL ES 2.0. Some of the code assumed iOS or Android, so
this needed to be fixed.

It appeared link_libraries was giving me problems early on, so I changed
to target_link_libraries. I'm not sure if this change is still needed,
   but link_libraries is deprecated, so it's probably a good change
   regardless.

There seems to be a bug in the Pi libraries for glBlendEquationOES,
	  glBlendEquationSeparateOES, glBlendFuncSpearateOES. Despite the
	  headers, these symbols are missing in the library, but exist
	  without the OES suffix. A hack to use the non-OES names is
	  employed.